requirements.txt
python-telegram-bot[job-queue]
requests
python-dotenv
web3
matplotlib
pandas
numpy
yfinance
textblob
